Renaming App to Main Separating components into history and progress directories
