+++ src/components/progress/progress-time-canvas.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import {setShadow, strokeCircle} from '../canvas.util';
+import {setShadow, strokeCircle} from '../../canvas.util';
 
 export default class ProgressTimeCanvas extends Component {
   componentDidMount() {
